+++ src/main/java/ru/yandex/money/gradle/plugins/gradleproject/ExtensionConfigurator.java
@@ -3,6 +3,7 @@ package ru.yandex.money.gradle.plugins.gradleproject;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import org.apache.commons.lang3.StringUtils;
 import org.gradle.api.Project;
+import org.gradle.api.tasks.wrapper.Wrapper;
 import org.gradle.plugin.devel.GradlePluginDevelopmentExtension;
 import ru.yandex.money.gradle.plugin.architecturetest.ArchitectureTestExtension;
 import ru.yandex.money.gradle.plugins.gradleproject.git.GitManager;
@@ -10,9 +11,15 @@ import ru.yandex.money.gradle.plugins.javapublishing.JavaArtifactPublishExtensio
 import ru.yandex.money.gradle.plugins.javapublishing.JavaArtifactPublishPlugin;
 import ru.yandex.money.gradle.plugins.library.git.expired.branch.settings.EmailConnectionExtension;
 import ru.yandex.money.gradle.plugins.library.git.expired.branch.settings.GitConnectionExtension;
+import ru.yoomoney.gradle.plugins.backend.build.JavaExtension;
+import ru.yoomoney.gradle.plugins.library.dependencies.CheckDependenciesPluginExtension;
+import ru.yoomoney.gradle.plugins.library.dependencies.checkversion.MajorVersionCheckerExtension;
 import ru.yoomoney.gradle.plugins.release.ReleaseExtension;
 
 import java.util.Arrays;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
 
 /**
  * Конфигуратор настроек плагинов.
@@ -33,11 +40,37 @@ public class ExtensionConfigurator {
     static void configure(Project project) {
         configureGitExpiredBranchesPlugin(project);
         configureReleasePlugin(project);
+        configureWrapper(project);
         configureArchitectureTestPlugin(project);
+        configureJavaPlugin(project);
+        configureCheckDependenciesExtension(project);
+        configureMajorVersionCheckerExtension(project);
+    }
+
+    private static void configureWrapper(Project project) {
+        project.getTasks().maybeCreate("wrapper", Wrapper.class)
+                .setDistributionUrl("https://nexus.yamoney.ru/content/repositories/" +
+                        "http-proxy-services.gradle.org/distributions/gradle-6.0.1-all.zip");
+    }
+
+    private static void configureJavaPlugin(Project project) {
+        List<String> repositories = List.of(
+                "https://nexus.yamoney.ru/content/repositories/releases/",
+                "https://nexus.yamoney.ru/content/repositories/jcenter.bintray.com/",
+                "https://nexus.yamoney.ru/content/repositories/thirdparty/",
+                "https://nexus.yamoney.ru/content/repositories/central/");
+
+        List<String> snapshotsRepositories = List.of(
+                project.getRepositories().mavenLocal().getUrl().toString(),
+                "https://nexus.yamoney.ru/content/repositories/snapshots/");
+
+        JavaExtension extension = project.getExtensions().getByType(JavaExtension.class);
+        extension.setRepositories(repositories);
+        extension.setSnapshotsRepositories(snapshotsRepositories);
     }
 
     private static String getStringExtProperty(Project project, String propertyName) {
-        String value = (String)project.getExtensions().getExtraProperties().get(propertyName);
+        String value = (String) project.getExtensions().getExtraProperties().get(propertyName);
         if (StringUtils.isBlank(value)) {
             throw new IllegalArgumentException("property " + propertyName + " is empty");
         }
@@ -103,5 +136,31 @@ public class ExtensionConfigurator {
     private static void configureArchitectureTestPlugin(Project project) {
         ArchitectureTestExtension architectureTestExtension = project.getExtensions().getByType(ArchitectureTestExtension.class);
         architectureTestExtension.getInclude().add("check_unique_enums_codes");
+    }
+
+    private static void configureMajorVersionCheckerExtension(Project project) {
+        Set<String> includeGroupId = new HashSet<>();
+        includeGroupId.add("ru.yamoney");
+        includeGroupId.add("ru.yandex.money");
+        includeGroupId.add("ru.yoomoney");
+
+        ru.yoomoney.gradle.plugins.backend.build.git.GitManager gitManager =
+                new ru.yoomoney.gradle.plugins.backend.build.git.GitManager(project);
+
+        MajorVersionCheckerExtension extension = project.getExtensions().getByType(MajorVersionCheckerExtension.class);
+
+        extension.includeGroupIdPrefixes = includeGroupId;
+
+        extension.failBuild = gitManager.isDevelopmentBranch();
+    }
+
+    private static void configureCheckDependenciesExtension(Project project) {
+        CheckDependenciesPluginExtension checkDependenciesPluginExtension = project.getExtensions()
+                .getByType(CheckDependenciesPluginExtension.class);
+
+        checkDependenciesPluginExtension.exclusionsRulesSources = List.of(
+                "ru.yandex.money.platform:yamoney-libraries-dependencies",
+                "libraries-versions-exclusions.properties"
+        );
     }
 }