+++ src/main/java/ru/yandex/money/gradle/plugins/gradleproject/GradleProjectPlugin.java
@@ -1,14 +1,16 @@
 package ru.yandex.money.gradle.plugins.gradleproject;
 
+import io.spring.gradle.dependencymanagement.DependencyManagementPlugin;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
 import org.gradle.plugin.devel.plugins.JavaGradlePluginPlugin;
 import org.gradle.util.VersionNumber;
 import ru.yandex.money.gradle.plugin.architecturetest.ArchitectureTestPlugin;
-import ru.yandex.money.gradle.plugins.backend.build.JavaModulePlugin;
 import ru.yandex.money.gradle.plugins.javapublishing.JavaArtifactPublishPlugin;
 import ru.yandex.money.gradle.plugins.library.git.expired.branch.GitExpiredBranchPlugin;
 import ru.yandex.money.gradle.plugins.task.monitoring.BuildMonitoringPlugin;
+import ru.yoomoney.gradle.plugins.backend.build.JavaPlugin;
+import ru.yoomoney.gradle.plugins.library.dependencies.CheckDependenciesPlugin;
 import ru.yoomoney.gradle.plugins.release.ReleasePlugin;
 
 import java.util.Arrays;
@@ -23,7 +25,9 @@ import java.util.Collection;
 public class GradleProjectPlugin implements Plugin<Project> {
 
     private static final Collection<Class<?>> PLUGINS_TO_APPLY = Arrays.asList(
-            JavaModulePlugin.class,
+            JavaPlugin.class,
+            DependencyManagementPlugin.class,
+            CheckDependenciesPlugin.class,
             JavaArtifactPublishPlugin.class,
             ReleasePlugin.class,
             GitExpiredBranchPlugin.class,